+++ packages/medusa/src/api/routes/store/customers/create-customer.ts
@@ -51,10 +51,7 @@ import { validator } from "../../../../utils/validator"
  *     content:
  *       application/json:
  *         schema:
- *           type: object
- *           properties:
- *             customer:
- *               $ref: "#/components/schemas/Customer"
+ *           $ref: "#/components/schemas/StoreCustomersRes"
  *   422:
  *     description: A customer with the same email exists
  *     content: