+++ packages/medusa/src/api/routes/store/orders/index.ts
@@ -83,6 +83,13 @@ export const defaultStoreOrdersFields = [
   "total",
 ] as (keyof Order)[]
 
+/**
+ * @schema StoreOrdersRes
+ * type: object
+ * properties:
+ *   order:
+ *     $ref: "#/components/schemas/Order"
+ */
 export type StoreOrdersRes = {
   order: Order
 }