+++ packages/medusa/src/api/routes/store/regions/list-regions.ts
@@ -80,12 +80,7 @@ import { validator } from "../../../../utils/validator"
  *     content:
  *       application/json:
  *         schema:
- *           type: object
- *           properties:
- *             regions:
- *               type: array
- *               items:
- *                 $ref: "#/components/schemas/Region"
+ *           $ref: "#/components/schemas/StoreRegionsListRes"
  *   "400":
  *     $ref: "#/components/responses/400_error"
  *   "404":