+++ packages/medusa/src/api/routes/store/returns/index.ts
@@ -12,6 +12,13 @@ export default (app) => {
   return app
 }
 
+/**
+ * @schema StoreReturnsRes
+ * type: object
+ * properties:
+ *   return:
+ *     $ref: "#/components/schemas/Return"
+ */
 export type StoreReturnsRes = {
   return: Return
 }